smart_home_api/requirements.txt
fastapi
uvicorn
paho-mqtt 